requirements.txt
streamlit
transformers==4.30.2  # Example:  Specifying a version
torch  # Example: you may need to include it.
google-generativeai